add basic class for TestRunListener which use TestRunResult
